check if skule-vote exists when bringing down

+++ scripts/application_stop.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
-cd /home/ec2-user/skule-vote
-docker-compose -f deployment/docker-compose.prod.yml down
+
+FOLDER=/home/ec2-user/skule-vote
+if [ -e "$FOLDER" ]; then
+  cd $FOLDER
+  docker-compose -f deployment/docker-compose.prod.yml down
+else
+  echo "$FOLDER does not exist."
+fi
